+++ src/App.js
@@ -1,14 +1,13 @@
-import React, { Component } from 'react';
-import { BrowserRouter as Router } from "react-router-dom";
-import RouterURL from './router/rerouter';
-
+import React, {Component} from 'react';
+import {BrowserRouter as Router} from 'react-router-dom';
+import RouterURL from './router/router';
 
 class App extends Component {
   render() {
     return (
       <Router>
         <div>
-          <RouterURL/>
+          <RouterURL />
         </div>
       </Router>
     );
